#5 refactor: 폴더명 단수형으로 수정 및 path 변경
