+++ src/mock/handler/index.ts
@@ -1,0 +1,3 @@
+import { showCaseHandler } from '@/mock/handler/showCase';
+
+export const handlers = [...showCaseHandler];
