+++ src/mock/handler/showCase.ts
@@ -1,5 +1,5 @@
 import { HttpResponse, http } from 'msw';
 
-import { showCaseData } from '@/mocks/data/showCase';
+import { showCaseData } from '@/mock/data/showCase';
 
 export const showCaseHandler = [http.get('https://msw-test.com', () => HttpResponse.json(showCaseData))];
