--- src/mocks/handler/index.ts
@@ -1,3 +1,0 @@
-import { showCaseHandler } from '@/mocks/handler/showCase';
-
-export const handlers = [...showCaseHandler];
